Agregando esquema inicial correcto de la base de datos

+++ backend/tufinca/src/main/java/com/gestion/tufinca/models/AdministradorMensajeId.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @NoArgsConstructor   // Constructor sin argumentos
 @Builder
 public class AdministradorMensajeId implements Serializable {
-    private Integer administrador;
-    private Integer mensaje;
+    private Long administrador;
+    private Long mensaje;
 }
